fix new channel bug.
Summary: refactor and fix bug where color never gets assigned if there's only 1 update_history. different than the problem we're seeing on production. one of the labels for an update_history entry is improperly labelled {F6524}. We should update to the latest version of sitrep when we can and it'll resolve this problem.

Test Plan: eye test.

Reviewers: cbrown

Reviewed By: cbrown

Differential Revision: https://phabricator.btlive.tv/D2221